+++ code/CQRSExtension.php
@@ -1,15 +1,29 @@
 <?php
 
+/**
+ * Class CQRSExtension
+ */
 class CQRSExtension extends Extension {
 
     private static $better_buttons_actions = [
         'writeToPayloadStore'
     ];
 
+    /**
+     * @var string
+     */
     private $key;
 
+    /**
+     * @var PayloadManifestParser
+     */
     private $parser;
 
+    /**
+     * CQRSExtension constructor.
+     *
+     * @param $key string
+     */
     public function __construct($key) {
         parent::__construct();
 
@@ -17,25 +31,48 @@ class CQRSExtension extends Extension {
         $this->parser = new PayloadManifestParser();
     }
 
+    /**
+     * @return PayloadStoreHandler
+     */
     public function getActiveHandler() {
         return RedisPayloadStoreHandler::inst();
     }
 
+    /**
+     * @return string
+     */
     private function getPayloadStoreKey() {
         return $this->owner->{$this->key} ?: $this->key;
     }
 
+    /**
+     * Obtains the commited payload
+     *
+     * @return array
+     */
     private function getCommitedPayload() {
         return $this->getActiveHandler()->read($this->getPayloadStoreKey());
     }
 
+    /**
+     * Generates a MD5 hash of the given payload.
+     *
+     * @param $payload
+     *
+     * @return string
+     */
     private function getPayloadChecksum($payload) {
         return md5(Convert::array2json($payload));
     }
 
+    /**
+     * Checks if the current data is equal to the commited payload.
+     *
+     * @return bool
+     */
     public function isInSync() {
         return $this->getPayloadChecksum($this->getCommitedPayload()) ===
-            $this->getPayloadChecksum($this->owner->commit());
+            $this->getPayloadChecksum($this->parser->commit($this->owner, false));
     }
 
     public function updateCMSActions($actions) {
@@ -46,13 +83,26 @@ class CQRSExtension extends Extension {
                 ->addExtraClass('ss-ui-action-constructive')->setAttribute('data-icon', 'accept'));
 
             // Add action for writing data to payload store
-            // TODO: Check for classes and icon
             if (Permission::check('PUBLISH_' . mb_strtoupper($this->owner->class))) {
-                $actions->push(FormAction::create('writeToPayloadStore', _t('CQRSExtension.WRITE', 'Lesedatenbank aktualisieren')));
+                $updateAction = FormAction::create(
+                    'writeToPayloadStore',
+                    _t('CQRSExtension.WRITE', 'Lesedatenbank aktualisieren')
+                )->setDisabled($this->isInSync());
+
+                if ($this->isInSync()) {
+                    $updateAction->setDescription(_t('CQRSExtension.UP_TO_DATE', 'Lesedatenbank ist aktuell'));
+                }
+
+                $actions->push($updateAction);
             }
         }
     }
 
+    /**
+     * Writes the commited payload to the current payload store.
+     *
+     * @return bool successful or not
+     */
     public function writeToPayloadStore() {
         $payload = $this->parser->commit($this->owner);
 
@@ -72,13 +122,32 @@ class CQRSExtension extends Extension {
         }
     }
 
+    /**
+     * Provides BetterButtonsAction for Super-Admin View.
+     *
+     * TODO: Disabled state not working when in super admin view
+     *
+     * @param $actions
+     */
     public function updateBetterButtonsActions($actions) {
-        $actions->push(BetterButtonCustomAction::create(
+        $updateAction = BetterButtonCustomAction::create(
             'writeToPayloadStore',
             _t('CQRSExtension.WRITE', 'Lesedatenbank aktualisieren')
-        ));
+        )->setDisabled($this->isInSync());
+
+        if ($this->isInSync()) {
+            $updateAction->
+            $updateAction->setDescription(_t('CQRSExtension.UP_TO_DATE', 'Lesedatenbank ist aktuell'));
+        }
+
+        $actions->push($updateAction);
     }
 
+    /**
+     * Renders a list of validation errors.
+     *
+     * @return mixed
+     */
     public function getErrorMessageForTemplate() {
         return $this->owner->customise([
             'ValidationErrors' => ArrayList::create(array_map(function ($error) {