+++ code/KeyTransformer.php
@@ -1,6 +1,14 @@
 <?php
 
+/**
+ * Interface KeyTransformer
+ */
 interface KeyTransformer {
 
+    /**
+     * @param string $value
+     *
+     * @return string
+     */
     public static function transform(string $value): string;
 }