+++ CHANGELOG.md
@@ -3,7 +3,6 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
+- Config over Code implementation
 - Functionality for checking if read and write databases are in sync
-
-### Fixed
 - read() method in Redis handler